+++ hoot-core/src/main/cpp/hoot/core/scoring/GraphComparator.cpp
@@ -414,7 +414,7 @@ void GraphComparator::_exportGraphImage(OsmMapPtr map, DirectedGraph& /*graph*/,
     }
     else
     {
-      double v = (cost / maxCost) * 255.0;
+      int v = (int)((cost / maxCost) * 255.0);
       c.setRgb(255, v, 0);
     }
